fastapi_app/requirements.txt
fastapi==0.115.6
jinja2==3.1.5
boto3==1.34.149
python-dotenv==1.0.0
transformers==4.43.3
torch==2.3.1
uvicorn==0.34.0
python-multipart==0.0.18